backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.11
openai==1.31.0
python-dotenv==1.0.0
pandas==2.2.3
